FDtomo-C Version 1.0 (#3)
* Version 1.0

* implement glatinv

* type casting 避免warning

* 使用 return 修正 Visual C++ 無法去除空白的問題

* 將runlsqr與sphrayderv的header分類資料夾

* 重新命名檔案(移除 c_XXX)

* 修正 header

* 修復字串超界問題

* 無法收斂版本

* 修復越界時略過地震的讀檔問題

* 修復無法收斂問題 (typo)

* 使用 fwrite 撰寫 fp_dts 二進制檔

* 修改字串長度上限

* 修復邊界 與 typo

    - [x] 修復邊界問題
    - [x] fix some typo
    - [x] remove useless debug message

* 輸出二進制檔

* 修改temp字串長度

* sphrayderv's function

* Update CMakeLists.txt

+++ src/time_process.c
@@ -1,4 +1,4 @@
-#include "../include/c_time_process.h"
+#include "../include/time_process.h"
 
 int isleap(int year) {
 	int leap = 0;
@@ -9,54 +9,55 @@ int isleap(int year) {
 }
 
 void etoh(double epoch, int *iyear, int *iday, int *ihour, int *imin,
-		double *sec) {
+	double *sec) {
 	int dayinmon[13] = { 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31, 31 };
 	/*
 	 char monname[12][4] = { "Jan\0", "Feb\0", "Mar\0", "Apr\0", "May\0", "Jun\0",
 	 "Jul\0", "Aug\0", "Sep\0", "Oct\0", "Nov\0", "Dec\0"
 	 };
 	 */
-	int doy = epoch / 86400.0;
+	int doy = (int)(epoch / 86400.0);
 	double secleft = fmod(epoch, 86400.0);
 	int hour = 0;
 	int minute = 0;
-	float second = 0.0;
+	double second = 0.0;
 
-//---compute hours minutes seconds
+	//---compute hours minutes seconds
 	if (secleft != 0.0) {
-//---before 1970
+		//---before 1970
 		if (secleft < 0) {
-//---subtract a day
+			//---subtract a day
 			doy = doy - 1;
-//----add a day
+			//----add a day
 			secleft = secleft + 86400;
 		}
-		hour = secleft / 3600;
+		hour = (int)(secleft / 3600);
 		secleft = fmod(secleft, 3600.0);
-		minute = secleft / 60;
+		minute = (int)(secleft / 60);
 		second = fmod(secleft, 60.0);
 	}
 
 	int year, diy;
 	if (doy >= 0) {
 		year = 1970;
-		a5: diy = 365 + isleap(year);
+	a5: diy = 365 + isleap(year);
 		if (doy < diy) {
 			goto a10;
 		}
 		doy -= diy;
 		year++;
 		goto a5;
-	} else {
+	}
+	else {
 		year = 1969;
-		a7: diy = 365 + isleap(year);
+	a7: diy = 365 + isleap(year);
 		if (doy >= 0)
 			goto a10;
 		doy += diy;
 		year--;
 		goto a7;
 	}
-	a10: doy++;
+a10: doy++;
 	//int date = year * 1000 + doy;
 	int leap = isleap(year);
 	int day = doy;
@@ -96,12 +97,13 @@ void dtoepoch(int date, double *time) {
 			days += 365;
 			days += isleap(i);
 		}
-	} else if (year < 1970) {
+	}
+	else if (year < 1970) {
 		for (int i = year; i <= 1969; i++) {
 			days -= 365;
 			days -= isleap(i);
 		}
 	}
 	days += (day - 1);
-	*time = days * 24 * 60 * 60;
+	*time = (double)(days * 24 * 60 * 60);
 }
